requirements.txt
streamlit
pandas
plotly
python-dotenv
bcrypt
pillow
reportlab